+++ Jenkins iOS SampleTests/Jenkins_iOS_SampleTests.swift
@@ -21,7 +21,9 @@ class Jenkins_iOS_SampleTests: XCTestCase {
     func testFabricAPIKeyValidationRejectsPlaceholders() {
         XCTAssertFalse(isConfiguredFabricAPIKey("$(FABRIC_API_KEY)"))
         XCTAssertFalse(isConfiguredFabricAPIKey("YOUR_FABRIC_API_KEY"))
+        XCTAssertFalse(isConfiguredFabricAPIKey("your_fabric_api_key"))
         XCTAssertFalse(isConfiguredFabricAPIKey("REPLACE_WITH_FABRIC_API_KEY"))
+        XCTAssertFalse(isConfiguredFabricAPIKey("replace_with_fabric_api_key"))
     }
 
     func testFabricAPIKeyValidationAcceptsTrimmedRealValues() {